Preparatory work to make GeoJSON driver optional
