+++ ogr/ogrsf_frmts/carto/ogrcartodatasource.cpp
@@ -28,7 +28,7 @@
 
 #include "ogr_carto.h"
 #include "ogr_pgdump.h"
-#include "ogrgeojsonreader.h"
+#include "ogrlibjsonutils.h"
 
 /************************************************************************/
 /*                        OGRCARTODataSource()                        */
