+++ ogr/ogrsf_frmts/carto/ogrcartolayer.cpp
@@ -28,7 +28,7 @@
 
 #include "ogr_carto.h"
 #include "ogr_p.h"
-#include "ogrgeojsonreader.h"
+#include "ogrlibjsonutils.h"
 
 /************************************************************************/
 /*                         OGRCARTOLayer()                            */
