+++ ogr/ogrsf_frmts/geojson/ogrgeojsonutils.cpp
@@ -1030,163 +1030,6 @@ GeoJSONSourceType JSONFGDriverGetSourceType(GDALOpenInfo *poOpenInfo)
 }
 
 /************************************************************************/
-/*                           GeoJSONPropertyToFieldType()               */
-/************************************************************************/
-
-constexpr GIntBig MY_INT64_MAX = (((GIntBig)0x7FFFFFFF) << 32) | 0xFFFFFFFF;
-constexpr GIntBig MY_INT64_MIN = ((GIntBig)0x80000000) << 32;
-
-OGRFieldType GeoJSONPropertyToFieldType(json_object *poObject,
-                                        OGRFieldSubType &eSubType,
-                                        bool bArrayAsString)
-{
-    eSubType = OFSTNone;
-
-    if (poObject == nullptr)
-    {
-        return OFTString;
-    }
-
-    json_type type = json_object_get_type(poObject);
-
-    if (json_type_boolean == type)
-    {
-        eSubType = OFSTBoolean;
-        return OFTInteger;
-    }
-    else if (json_type_double == type)
-        return OFTReal;
-    else if (json_type_int == type)
-    {
-        GIntBig nVal = json_object_get_int64(poObject);
-        if (!CPL_INT64_FITS_ON_INT32(nVal))
-        {
-            if (nVal == MY_INT64_MIN || nVal == MY_INT64_MAX)
-            {
-                static bool bWarned = false;
-                if (!bWarned)
-                {
-                    bWarned = true;
-                    CPLError(
-                        CE_Warning, CPLE_AppDefined,
-                        "Integer values probably ranging out of 64bit integer "
-                        "range have been found. Will be clamped to "
-                        "INT64_MIN/INT64_MAX");
-                }
-            }
-            return OFTInteger64;
-        }
-        else
-        {
-            return OFTInteger;
-        }
-    }
-    else if (json_type_string == type)
-        return OFTString;
-    else if (json_type_array == type)
-    {
-        if (bArrayAsString)
-        {
-            eSubType = OFSTJSON;
-            return OFTString;
-        }
-        const auto nSize = json_object_array_length(poObject);
-        if (nSize == 0)
-        {
-            eSubType = OFSTJSON;
-            return OFTString;
-        }
-        OGRFieldType eType = OFTIntegerList;
-        for (auto i = decltype(nSize){0}; i < nSize; i++)
-        {
-            json_object *poRow = json_object_array_get_idx(poObject, i);
-            if (poRow != nullptr)
-            {
-                type = json_object_get_type(poRow);
-                if (type == json_type_string)
-                {
-                    if (i == 0 || eType == OFTStringList)
-                    {
-                        eType = OFTStringList;
-                    }
-                    else
-                    {
-                        eSubType = OFSTJSON;
-                        return OFTString;
-                    }
-                }
-                else if (type == json_type_double)
-                {
-                    if (eSubType == OFSTNone &&
-                        (i == 0 || eType == OFTRealList ||
-                         eType == OFTIntegerList || eType == OFTInteger64List))
-                    {
-                        eType = OFTRealList;
-                    }
-                    else
-                    {
-                        eSubType = OFSTJSON;
-                        return OFTString;
-                    }
-                }
-                else if (type == json_type_int)
-                {
-                    if (eSubType == OFSTNone && eType == OFTIntegerList)
-                    {
-                        GIntBig nVal = json_object_get_int64(poRow);
-                        if (!CPL_INT64_FITS_ON_INT32(nVal))
-                            eType = OFTInteger64List;
-                    }
-                    else if (eSubType == OFSTNone &&
-                             (eType == OFTInteger64List ||
-                              eType == OFTRealList))
-                    {
-                        // ok
-                    }
-                    else
-                    {
-                        eSubType = OFSTJSON;
-                        return OFTString;
-                    }
-                }
-                else if (type == json_type_boolean)
-                {
-                    if (i == 0 ||
-                        (eType == OFTIntegerList && eSubType == OFSTBoolean))
-                    {
-                        eSubType = OFSTBoolean;
-                    }
-                    else
-                    {
-                        eSubType = OFSTJSON;
-                        return OFTString;
-                    }
-                }
-                else
-                {
-                    eSubType = OFSTJSON;
-                    return OFTString;
-                }
-            }
-            else
-            {
-                eSubType = OFSTJSON;
-                return OFTString;
-            }
-        }
-
-        return eType;
-    }
-    else if (json_type_object == type)
-    {
-        eSubType = OFSTJSON;
-        return OFTString;
-    }
-
-    return OFTString;  // null
-}
-
-/************************************************************************/
 /*                        GeoJSONStringPropertyToFieldType()            */
 /************************************************************************/
 
@@ -1220,32 +1063,4 @@ OGRFieldType GeoJSONStringPropertyToFieldType(json_object *poObject,
         // TODO: What if both are false?
     }
     return OFTString;
-}
-
-/************************************************************************/
-/*                           OGRGeoJSONGetGeometryName()                */
-/************************************************************************/
-
-const char *OGRGeoJSONGetGeometryName(OGRGeometry const *poGeometry)
-{
-    CPLAssert(nullptr != poGeometry);
-
-    const OGRwkbGeometryType eType = wkbFlatten(poGeometry->getGeometryType());
-
-    if (wkbPoint == eType)
-        return "Point";
-    else if (wkbLineString == eType)
-        return "LineString";
-    else if (wkbPolygon == eType)
-        return "Polygon";
-    else if (wkbMultiPoint == eType)
-        return "MultiPoint";
-    else if (wkbMultiLineString == eType)
-        return "MultiLineString";
-    else if (wkbMultiPolygon == eType)
-        return "MultiPolygon";
-    else if (wkbGeometryCollection == eType)
-        return "GeometryCollection";
-
-    return "Unknown";
 }
