+++ ogr/ogrsf_frmts/geojson/ogrtopojsonreader.cpp
@@ -28,6 +28,7 @@
 
 #include "ogrgeojsonreader.h"
 #include "ogrgeojsonutils.h"
+#include "ogrlibjsonutils.h"
 #include "ogr_geojson.h"
 #include <json.h>  // JSON-C
 #include "ogr_api.h"
