+++ ogr/ogrsf_frmts/plscenes/CMakeLists.txt
@@ -22,7 +22,6 @@ set_property(
   PROPERTY RESOURCE "${GDAL_DATA_FILES}")
 
 gdal_standard_includes(ogr_PLSCENES)
-target_include_directories(ogr_PLSCENES PRIVATE $<TARGET_PROPERTY:ogr_geojson,SOURCE_DIR>)
 
 if (GDAL_USE_JSONC_INTERNAL)
   gdal_add_vendored_lib(ogr_PLSCENES libjson)
