+++ ogr/ogrsf_frmts/plscenes/ogrplscenesdatav1dataset.cpp
@@ -27,7 +27,7 @@
  ****************************************************************************/
 
 #include "ogr_plscenes.h"
-#include "ogrgeojsonreader.h"
+#include "ogrlibjsonutils.h"
 #include <time.h>
 
 /************************************************************************/
